+++ vscode-extension/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "wireit" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+- Use the same logic as the CLI for finding diagnostics.
+
 ## [0.2.0] - 2022-05-04
 
 - Add code actions to fix some common mistakes, as well as to convert a script